Implement Rare Feature Encoder (#546)
* Fix expected value in Baseclass.rst
Changed RandomForest test from doctest::  to code-block:: python
The result changed between 82 and 83. 

* Implement Rare feature Encoder transformer
The Transformer is implemented in rare_feature.py
The Transformer is tested in test_transformers.py

* Update CHANGELOG.md
Added explanation and example of the new Transformer. 

* Code optimization
The Rare Feature Encoder code is optimized

* Updated changes
The recommended changes have been updated

Co-authored-by: Iben Fjord Kjærsgaard <abibkj@almbrand.dk>
Co-authored-by: Anders Bogsnes <andersbogsnes@gmail.com>